5- Airline Reservation System On ASP.Net MVC_ Admin Panel Part II(720P_HD)

+++ ARS/Controllers/AeroPlaneController.cs
@@ -18,28 +18,51 @@ namespace ARS.Controllers
         // GET: AeroPlane
         public ActionResult Index()
         {
-            return View(db.PlaneInfo.ToList());
+            if (Session["admin"] != null)
+            {
+                return View(db.PlaneInfo.ToList());
+            }
+            else
+            {
+                return RedirectToAction("AdminLogin", "Admin");
+            }
+            
         }
 
         // GET: AeroPlane/Details/5
         public ActionResult Details(int? id)
         {
-            if (id == null)
+            if (Session["admin"] != null)
             {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                if (id == null)
+                {
+                    return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                }
+                AeroPlaneInfo aeroPlaneInfo = db.PlaneInfo.Find(id);
+                if (aeroPlaneInfo == null)
+                {
+                    return HttpNotFound();
+                }
+                return View(aeroPlaneInfo);
             }
-            AeroPlaneInfo aeroPlaneInfo = db.PlaneInfo.Find(id);
-            if (aeroPlaneInfo == null)
+            else
             {
-                return HttpNotFound();
+                return RedirectToAction("AdminLogin", "Admin");
             }
-            return View(aeroPlaneInfo);
         }
 
         // GET: AeroPlane/Create
         public ActionResult Create()
         {
-            return View();
+            if (Session["admin"] != null)
+            {
+                return View();
+            }
+            else
+            {
+                return RedirectToAction("AdminLogin", "Admin");
+            }
+            
         }
 
         // POST: AeroPlane/Create
@@ -64,16 +87,24 @@ namespace ARS.Controllers
         // GET: AeroPlane/Edit/5
         public ActionResult Edit(int? id)
         {
-            if (id == null)
+            if (Session["admin"] != null)
             {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                if (id == null)
+                {
+                    return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                }
+                AeroPlaneInfo aeroPlaneInfo = db.PlaneInfo.Find(id);
+                if (aeroPlaneInfo == null)
+                {
+                    return HttpNotFound();
+                }
+                return View(aeroPlaneInfo);
             }
-            AeroPlaneInfo aeroPlaneInfo = db.PlaneInfo.Find(id);
-            if (aeroPlaneInfo == null)
+            else
             {
-                return HttpNotFound();
+                return RedirectToAction("AdminLogin", "Admin");
             }
-            return View(aeroPlaneInfo);
+            
         }
 
         // POST: AeroPlane/Edit/5
@@ -95,16 +126,27 @@ namespace ARS.Controllers
         // GET: AeroPlane/Delete/5
         public ActionResult Delete(int? id)
         {
-            if (id == null)
+
+            if (Session["admin"] != null)
             {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                if (id == null)
+                {
+                    return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                }
+
+                AeroPlaneInfo aeroPlaneInfo = db.PlaneInfo.Find(id);
+
+                if (aeroPlaneInfo == null)
+                {
+                    return HttpNotFound();
+                }
+                return View(aeroPlaneInfo);
             }
-            AeroPlaneInfo aeroPlaneInfo = db.PlaneInfo.Find(id);
-            if (aeroPlaneInfo == null)
+            else
             {
-                return HttpNotFound();
+                return RedirectToAction("AdminLogin", "Admin");
             }
-            return View(aeroPlaneInfo);
+            
         }
 
         // POST: AeroPlane/Delete/5
@@ -112,10 +154,17 @@ namespace ARS.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult DeleteConfirmed(int id)
         {
-            AeroPlaneInfo aeroPlaneInfo = db.PlaneInfo.Find(id);
-            db.PlaneInfo.Remove(aeroPlaneInfo);
-            db.SaveChanges();
-            return RedirectToAction("Index");
+            if (Session["admin"] != null)
+            {
+                AeroPlaneInfo aeroPlaneInfo = db.PlaneInfo.Find(id);
+                db.PlaneInfo.Remove(aeroPlaneInfo);
+                db.SaveChanges();
+                return RedirectToAction("Index");
+            }
+            else
+            {
+                return RedirectToAction("AdminLogin", "Admin");
+            }
         }
 
         protected override void Dispose(bool disposing)
