+++ ARS/Controllers/FlightBookingsController.cs
@@ -17,30 +17,54 @@ namespace ARS.Controllers
         // GET: FlightBookings
         public ActionResult Index()
         {
-            var flightsBookings = db.FlightsBookings.Include(f => f.TicketResevation_tbls);
-            return View(flightsBookings.ToList());
+            if (Session["admin"] != null)
+            {
+                var flightsBookings = db.FlightsBookings.Include(f => f.TicketResevation_tbls);
+                return View(flightsBookings.ToList());
+            }
+            else
+            {
+                return RedirectToAction("AdminLogin", "Admin");
+            }
+            
         }
 
         // GET: FlightBookings/Details/5
         public ActionResult Details(int? id)
         {
-            if (id == null)
+            if (Session["admin"] != null)
             {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                if (id == null)
+                {
+                    return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                }
+                FlightBooking flightBooking = db.FlightsBookings.Find(id);
+                if (flightBooking == null)
+                {
+                    return HttpNotFound();
+                }
+                return View(flightBooking);
             }
-            FlightBooking flightBooking = db.FlightsBookings.Find(id);
-            if (flightBooking == null)
+            else
             {
-                return HttpNotFound();
+                return RedirectToAction("AdminLogin", "Admin");
             }
-            return View(flightBooking);
+            
         }
 
         // GET: FlightBookings/Create
         public ActionResult Create()
         {
-            ViewBag.ResId = new SelectList(db.TicketReservation_tbl, "ResId", "RestFrom");
-            return View();
+            if (Session["admin"] != null)
+            {
+                ViewBag.ResId = new SelectList(db.TicketReservation_tbl, "ResId", "RestFrom");
+                return View();
+            }
+            else
+            {
+                return RedirectToAction("AdminLogin", "Admin");
+            }
+            
         }
 
         // POST: FlightBookings/Create
@@ -50,31 +74,47 @@ namespace ARS.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Create([Bind(Include = "bId,bCusName,To,bCusEmail,bCusSeats,bCusPhone,bCusCnic,ResId")] FlightBooking flightBooking)
         {
-            if (ModelState.IsValid)
+            if (Session["admin"] != null)
             {
-                db.FlightsBookings.Add(flightBooking);
-                db.SaveChanges();
-                return RedirectToAction("Index");
-            }
+                if (ModelState.IsValid)
+                {
+                    db.FlightsBookings.Add(flightBooking);
+                    db.SaveChanges();
+                    return RedirectToAction("Index");
+                }
 
-            ViewBag.ResId = new SelectList(db.TicketReservation_tbl, "ResId", "RestFrom", flightBooking.ResId);
-            return View(flightBooking);
+                ViewBag.ResId = new SelectList(db.TicketReservation_tbl, "ResId", "RestFrom", flightBooking.ResId);
+                return View(flightBooking);
+            }
+            else
+            {
+                return RedirectToAction("AdminLogin", "Admin");
+            }
+            
         }
 
         // GET: FlightBookings/Edit/5
         public ActionResult Edit(int? id)
         {
-            if (id == null)
+            if (Session["admin"] != null)
             {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                if (id == null)
+                {
+                    return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                }
+                FlightBooking flightBooking = db.FlightsBookings.Find(id);
+                if (flightBooking == null)
+                {
+                    return HttpNotFound();
+                }
+                ViewBag.ResId = new SelectList(db.TicketReservation_tbl, "ResId", "RestFrom", flightBooking.ResId);
+                return View(flightBooking);
             }
-            FlightBooking flightBooking = db.FlightsBookings.Find(id);
-            if (flightBooking == null)
+            else
             {
-                return HttpNotFound();
+                return RedirectToAction("AdminLogin", "Admin");
             }
-            ViewBag.ResId = new SelectList(db.TicketReservation_tbl, "ResId", "RestFrom", flightBooking.ResId);
-            return View(flightBooking);
+            
         }
 
         // POST: FlightBookings/Edit/5
@@ -84,29 +124,45 @@ namespace ARS.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Edit([Bind(Include = "bId,bCusName,To,bCusEmail,bCusSeats,bCusPhone,bCusCnic,ResId")] FlightBooking flightBooking)
         {
-            if (ModelState.IsValid)
+            if (Session["admin"] != null)
             {
-                db.Entry(flightBooking).State = EntityState.Modified;
-                db.SaveChanges();
-                return RedirectToAction("Index");
+                if (ModelState.IsValid)
+                {
+                    db.Entry(flightBooking).State = EntityState.Modified;
+                    db.SaveChanges();
+                    return RedirectToAction("Index");
+                }
+                ViewBag.ResId = new SelectList(db.TicketReservation_tbl, "ResId", "RestFrom", flightBooking.ResId);
+                return View(flightBooking);
             }
-            ViewBag.ResId = new SelectList(db.TicketReservation_tbl, "ResId", "RestFrom", flightBooking.ResId);
-            return View(flightBooking);
+            else
+            {
+                return RedirectToAction("AdminLogin", "Admin");
+            }
+            
         }
 
         // GET: FlightBookings/Delete/5
         public ActionResult Delete(int? id)
         {
-            if (id == null)
+            if (Session["admin"] != null)
             {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                if (id == null)
+                {
+                    return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                }
+                FlightBooking flightBooking = db.FlightsBookings.Find(id);
+                if (flightBooking == null)
+                {
+                    return HttpNotFound();
+                }
+                return View(flightBooking);
             }
-            FlightBooking flightBooking = db.FlightsBookings.Find(id);
-            if (flightBooking == null)
+            else
             {
-                return HttpNotFound();
+                return RedirectToAction("AdminLogin", "Admin");
             }
-            return View(flightBooking);
+            
         }
 
         // POST: FlightBookings/Delete/5
@@ -114,10 +170,18 @@ namespace ARS.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult DeleteConfirmed(int id)
         {
-            FlightBooking flightBooking = db.FlightsBookings.Find(id);
-            db.FlightsBookings.Remove(flightBooking);
-            db.SaveChanges();
-            return RedirectToAction("Index");
+            if (Session["admin"] != null)
+            {
+                FlightBooking flightBooking = db.FlightsBookings.Find(id);
+                db.FlightsBookings.Remove(flightBooking);
+                db.SaveChanges();
+                return RedirectToAction("Index");
+            }
+            else
+            {
+                return RedirectToAction("AdminLogin", "Admin");
+            }
+            
         }
 
         protected override void Dispose(bool disposing)
